Feature/social messaging bots (#147)
Our finding was that there is a misunderstanding of how to use Conversation Metadata vs. Structured Content metadata. IE the difference was conflated.
Secondly we actually do not look to have a straight forward example of conversation metadata in our SDK examples. So whilst, this is somewhat elaborate it serves a practical example.